requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
pillow>=8.3.0
albumentations>=1.1.0
dvc>=3.0.0
dvc[s3]>=3.0.0
pytest>=7.0.0
python-dotenv>=0.19.0
pyyaml>=5.4.1
timm>=0.6.0